First import commit. A lot of squashing.
This started as an internal script that would update a number of working
dirs for several projects with a new version of Ruby, in many files.
This has the magic "figure out the next version" method**, plus some more
routine stuff, and none of it focused on Gemfile specifically. I've been
re-working it for a while and am pushing it to this repo now.

A little worried that trying to share Ruby version updating with Gemfile
specifics isn't really worth it, Gemfile has its own needs.

** The original one was working way too hard, the current
calc_new_version just sorts strings. This makes a lot more sense. I
thought of it while trying out Gem::Version, knowing that would only
work for Gems and not Ruby versions starting with words ... then
realized it'd probably just work with straight-up strings, and --
wha-lah.

0.2.0 supports custom advisory db.

No more Gemfile.lock updating. It used to, to try and hack around
`bundle update` greedy behavior, but it doesn't really work reliably and
is pointless in many cases.