First commit of internal script.
I had an internal script that would update a number of working dirs for
several projects with a new version of Ruby, in many files. This has the
magic "figure out the next version" method, plus some more routine
stuff, and none of it focused on Gemfile specifically. I've started
re-working it, see the need for some more refactoring, but want to get
some feedback and am at a stopping point right now.