Refactorings
Specs are passing, but this is still an awkward place to stop.
Nevertheless, wanted to commit before moving on. This is starting to get
rid of the Updater class, and pushing that into UpdateSpec itself, as
well as having specific subclasses for updating the Ruby version and
then gems in a Gemfile.

This all feels a little wonky still, but I'm also just chipping in here
with a little free time so far. It'll get there.