Gemfile spec started, 1st spec passing & bugfix
To get the 1st spec passing in the new gemfile_spec file, found a bug in
the prior change on the gsub! block that was blowing away the whole
match, and due to sloppy matching in the rspec, I hadn't noticed.

Now that's fixed, and the first of many specs to come are passing.
Thanks big time to the RubyGems specifications for all the different
version variations - a few I never even knew about.