Special handling for >, >= cases added.
Rollin' along. Gemfile gettin' a little customized here, but still
enough shared stuff going on in UpdateSpec to justify it.

A little worried that trying to share Ruby version updating with Gemfile
specifics, when it gets to these custom pessimistic cases, will be too
forced. But, so far so good. More to come though.