Undo 'bugs' about major version upgrades.
I realized this tool shouldn't bump over major version upgrades as they'll
likely have breaking changes, so I rolled back those spec failures.

Except that exposed a funny case:

If the requested gem is `< 3`, the current version is `2.4` and the
patched is `2.5.1`, the Gemfile.lock should be updated, but the Gemfile
shouldn't be. So far there's been no need to really understand the
operator, but this case can force that, unfortunately, esp. since it's
probably not too common. This spec is still broken.