Conservative updates
The way this works is cool - the Molinillo dependency resolution engine
they built for bundler, is a generic engine, and takes orders from
whatever `Molinillo::SpecificationProvider` instance is passed to it -
which for bundler is `Bundler::Resolver`.

The `search_for` method in `Resolver` is part of that spec, and the docs
on the Molinillo module say to return that last in reverse order of
priority. What `search_for` returns is a sorted list of all available
gem versions - and the order matters, because that’s the order the
resolution engine will work in - which is why it grabs the most recent
version available to satisfy all of the requirement dependencies, as it
should.

In order to do this conservatively, we can (a) reverse the order (b)
remove _older_ versions so we don’t actually regress versions. The
engine is still doing the hard work of finding versions that satisfy
everything ... but going from oldest/current to most recent ... setting
up a point release of `foo` and perhaps no upgrade at all of
dependencies if the requirements are still met.