Change design from prepend module to extend style.
Prepending the module to the Resolver class modifies all instances, and
(for testing at least, but also for less intrusiveness) this approach
works better. It also is closer to what a patch to bundler itself would
need to look like.

An additional spec checks an update to two named gems.