Start supporting strict/minor allowed options.
I started testing the code in live 'update all gems conservatively'
situations and quickly hit some cases not being handled well. The first
was a Rails app locked requirements-wise to 3.2.22.2, but not locked on
the sass-rails gem. The next incremental version increase for sass-rails
was up to the next major version 4.0.0, which then led to an obvious
failure to resolve. In addition, no definition of conservative should
allow a major upgrade of a dependent gem ... and sometimes maybe not
even any gem down in the dependency tree.

Which led me to existing specs with notes on them about maybe this
should do this, and not that, forcing me to get more detailed on what
exactly "conservative" means.

I also had the thought that perhaps instead of the approach to actually
remove certain versions from the search_for results, which can lead to
unresolvable dependency trees, even though a version does exist in the
overall index which could satisfy it, that maybe it should just have a
more fine-tuned sorting order, without actually removing any versions
from the search_for results.

While exploring all of this, I've now settled on keeping both versions
(filter_specs which I'm now calling `--strict` mode), and a new approach
(sort_specs, the default) which will lead to fewer unresolvable
dependency trees, but more cases where an unexpected amount of upgrading
occurs, even some major version increases. I'm also adding a second
option for allowing minor_version increases, defaulting to just the most
recent release of the current major.minor version.

For example consider this case: If I'm ok with minor_version increases
(with the new option), and I currently have `foo (2.4.0)`, which depends
on `bar (>= 1.0.4)`, and I conservatively upgrade to `foo (2.5.0)` which
now depends on `bar (~> 2.0)` what should I do? Some definitions of
conservative would say "don't upgrade any gem up a major version" but
the code was told to conservatively upgrade `foo` ... and `foo` is only
going up one minor version.

In this case, `--strict` mode will cause this to fail, Bundler throwing
a `VersionConflict` exception because it won't be allowed to know about
any 2.x versions of `bar`. The default mode will allow this to happen.

This is far from resolved, and I don't have many cases covering the
various combinations, but this is a step in the right direction I feel,
and all but reminder specs are passing.

(Actually, the first real problem I hit was an as-yet not understood
failure of the resolution engine to untangle what otherwise was a fine
combination involving a JSON gem. I did experiment with removing my
memoization/hash-cache in search_for and that problem went away, so I
left it remarked out with a TODO and failing spec to come back to.)

Couple of other notes:

- Ensure proper handling of resolve_remotely! when "doing it live!" vs.
  test run.
- Ensure Bundler::UI::Shell is setup to read that output on remote
  resolve.