0.5.1 - bug fix for default (non-strict) mode.
If bundler is a requirement of a gem, Bundler itself keeps it out of the
proceedings, and there will be no locked_spec found ... the new
non-strict mode code didn't take that case into account. There's now
specs to cover this case for both non-strict and strict modes.

A couple of real life tests have shown this to be working, esp. better
in default mode.

A couple of problems at this point:

- strict mode demands the gems in question (all of them if none
  specified) be updated at least a release version or else it fails,
  when it could just stay put.

- there isn't a way to feed a list of gem names to the update cmd.

- nothing actually calls the Bundler code to install the updated gems.
  :)