0.5.1 con't - gem list passed in and installs.
I was worried that calling the Installer through the usual means as
Bundler does in its own update command wasn't going to work with the
customized Definition, but it is working so far. Which is cool :) And
awesome, because we're getting a fair amount of additional behavior
included for free - some sticky settings for example.

I'm not gathering all of the options, so that's a problem for now. And
the way Boson handles an array argument, we can't pass a list of gem
names into the command without putting that list in quotes, which is
unnatural.

But, what's awesome is I've run this on a few more real repos at LS
and it's working really well so far :)