0.5.2 - strict mode a little less strict.
Strict mode could blow up in some cases, because it would remove the
current version from the search_for result, attempting to force an
update to the next version. But if a suitable dependency tree couldn't
be found, Bundler would have to raise a VersionConflict. I coded it this
way before I realized sorting the results with a mix of descending by
release version first, then falling back to current versions was an
elegant way to prefer incrementing the version while allowing a fallback
to current.

Strict mode still filters out older and newer versions that bump major
version (and minor versions depending on that setting), but no longer
needs to filter out the current version, because it then sorts its
results through non-strict sorting. This accomplishes two things: (a)
Bundler doesn't have to raise a VersionConflict if staying on the
current version satisfies everything and (b) strict mode can go to the
LATEST release version, which is better and consistent with what
non-strict mode does.

It's debatable in my mind if anyone would want a strict_er_ mode that
only pushed a gem from 2.4.0 to 2.4.1 when 2.4.3 is available, I'm not
sure how that's that helpful, but it could conceivably done in the
future.