WIP: `patch` behavior not matching `update`.
Doing some experimental testing of the gem in for-realz life, I ran
across a few small things (listed below), but a bigger issue: if
patching version 1.4.4 due to a known vulnerability, and >= 1.4.5 is
patched, and 1.4.7 exists, `patch` still leaves the gem at 1.4.5, even
though `update` would go to the latest available release.

I decided to bite the bullet and add some integration testing around the
`patch` command, since I've never built it out. After running into a few
complications, figuring out that I couldn't easily continue to pass in a
faked out `Bundler::Definition` now that `conservative_update` is
running a full `bundle install` command, I decided to have a fake
Gemfile, fake advisory, but a real gem.

The test right now is failing, because it's taken me a few nights
hacking on the danged thing and now it's finally clear what the
disconnect is.

The prior behavior of the code patching a version for a vulnerability,
in the case where the Gemfile used to specify a specific version, was to
update it to the _most_ conservative patched version. The goal in that
case is "just get me patched, don't touch anything else."

After working for a while since then on a general conservative update,
in that mindset it makes little sense to bump up one release version
when a few more exist, all for bug fixes (presuming SemVer). That seemed
a better option in that case.

So - now how to reconcile the two mindsets? That I need to think on some
more, but I need to commit what I have so far.

Other changes:

- Fix: gem attr_reader rename to gem_name, since gem method already
  existed.
- Fix: duplicate vulnerable gems eliminated in `scan` output.
- --strict and --minor_allowed options added to `patch` command.
- _scan method reworked, new un-documented `:skip_bundler_advise` option
  added for testability.
- README Usage updated