Fix DEBUG_PATCH_RESOLVER output for all versions.
My prior clean-up kept it to only one version, but the whole point of
dumping this data is to see the full list of versions and the resulting
order (and dependencies, though that could get thick in some cases).

This was done to chase down a real case of not being able to patch
sprockets due to other parent gems up the tree being locked down. Now I
have better clarity on how to solve that problem.