Added passing spec on child/parent sec update.
Enhanced the PathedGemfileLockFixture to handle dependencies.

I _thought_ this would be failing. `tea` is the parent gem that's
locked, `nuu` is the child gem that needs to be updated for a security
patch. IIRC, what I'd seen in for realz was a case where the parent
wasn't upgrading because Bundler hadn't been given permission to unlock
it. So, now I'm stumped and need to go re-research to real life case,
but at least this is passing.