Support for patching lockfile only gems working.
Cool - this case is working now. The `patch` action now specifically
passes through the new_version we're targeting, and that is used to
jimmy up the locked_specs collection at such a time that it thinks
that's the version read in from the lockfile, and so things operate
accordingly from there.

This case got sort of left behind with all of the attention more
recently put on the general `update` command. A general conservative
update has no mandated version that needs to be obtained, just either
the most recent release or minor version.

`patch` for security vulnerabilities _does_ have a minimum mandated
version involved, which was fine thus far in cases where the gem and/or
version was specified in the Gemfile, but not in this case where the
vulnerable gem is only in the lockfile.

This should be a nice piece of functionality for the realz world.