The big, I think this is working, `patch` commit.
`update` doesn't really care about a _specific_ version to get to, it
gets to the most recent release (or minor with -m option) version.

`patch` is about getting secure, and getting to the nearest patched
version, a specific version, which may not be the most recent
release/minor version.

`patch` also needs to free all of the other specs to move around _if
they have to_ but otherwise stay put. Which is a different goal for a
general `update` command, which encourages _all_ gems to get up to the
most recent release/minor version.

But marking all gems "free to move around" is the same thing inside
Bundler, it's the `true` boolean being passed in as the `unlock`
parameter to the Definition.

But our conservative resolver needs to know the difference, because if
everything is unlocked, but in `patch` it wants to encourage everyone
(but the gems being patched) to _stay put_, and in `update` it wants to
encourage everyone to bump up, it's got to sort the versions
accordingly.

The problem before this change was Bundler was told _not_ to allow all
the other gems to move, and it would result in nothing happening (or
VersionConflict).

The other weird thing going on here, is to pass through the specific
version being sought for an insecure gem, is a Gem::Specification can go
through instead of just the gem name, in the list of gems_to_update.
That's used by faking out Bundler (which is always a dangerous design
plan) into being told the _currently_ locked version is the version we
want to get up to, so that in conjunction with our resolver, anything
prior to the closest patched version isn't even available (which is why
things can lead to VersionConflicts)

After many lost nights trying to figure out how to debug these real
world scenarios forcing these changes, and believing I could use Path
sources to test double remotes (this doesn't work - Path sources are
special and eagerly unlock things unlike remotes), this change finally
appears to be doing what I want, but I've cluttered the code with even
more TODOs and need to start cleaning up. So, 0.6.0 isn't ready for
release yet, but I can start heading there now.