Some crufty bits and one big glitch fix attempt.
So ... commit 806c9706 added a pretty good way at getting patching to be
more successful in freeing up the whole dependency tree to move without
causing any unnecessary moves.

But trying this out in a number of real life repositories led to two
problems.

The first problem was - strange - and very time consuming to track down,
and I haven't even fully gotten to the bottom of it. I tested it with a
repo that had vulnerable versions of sidekiq-pro and its dependency the
open source sidekiq, but it would end up in a weird place where
Bundler's index had _no_ versions of sidekiq in it, and would end in
VersionConflict. This was before it hit any of my ConservativeResolver
code. Compared to a full run of bundle update sidekiq-pro, which worked
fine, the only difference I could find was the sources inside the
Definition, had a rubygems_aggregrate with any empty @remotes ivar in
bundler-patch, but they had the proper array of sources. Patching the
Definition to make sure the @remote array was set fixed the problem, but
I can't recreate the problem in isolation (presumably because I need
multiple remotes in isolation and don't have fixtures to do that yet),
and there is potentially enough going on in all of the setup that
something is odd. Though I did replicate the problem with the same repo
by attempting a _full_ `bundle update` - it also cratered for the same
reason.

For now, I'm leaving the workaround in my code for this problem and will
move on.

The second problem is more common and understandable. There's a common
vulnerability in the `mail` gem that is used by `rails` and associates,
and there is no patched version of `mail` that is compatible with the
requirements by Rails 3.2.x. (The vulnerability is minor, so we're fine
with leaving it until we can upgrade into Rails 4). But this current
version of `bundler-patch` attempts to fix all other vulnerabilities at
the same time, and does so by faking out the locked_specs by setting it
to the version the vulnerable gem _should_ be on. This is fine as long
as that version is compatible, but if it's not, then when updating many
gems, the whole run fails with a VersionConflict, on a gem that can
never be compatible.

It would be better if we 'did our best' and allowed the one gem that's
'stuck' to stay put and get the rest patched. (It would be even better
to inform the user of this problem, but that's a fancier feature for
another day).

This commit doesn't have that fix in place at all, but that's what needs
to happen next.

The rest of the commit has some debugging enhancements and tweaks, but
mostly a change to allow the user to be warned of not being able to do
an upgrade of a vulnerability, most commonly because upgrades to a new
major version are not allowed.