All gems unlocked for better patching behavior.
This has the improvement in response to commit 806c9706.

GemPatch class introduced to properly capture the four pieces of data
that need to be grouped to pass that through so evaluation of patching
behavior can occur. This replaces the half-arsed usage of
Gem::Specification introduced in 806c9706.

Any patching attempt simulates an `update all`, but the sorting of
non-patching gems encourages it to stay put, but allows it to move
around if necessary to find a dependency tree that works. Sorting of
gems being patched is almost the same as an update of the gem, but
encourages just the closest patched version first. That behavior may go
away, why patch from 1.8.1 to 1.8.6 when 1.8.9 is out there too? Or
maybe that needs to be an option. I could see arguments either way. Or
maybe the strict option includes this behavior. Dunno.

This removes the ugly hacking of the locked_spec for patching gems,
which was 'clever' but also setup VersionConflict breakage as discussed
in more detail in the prior commit.

A couple of other incident things going on in this commit too:

- Better UI handling of gem patches output in Scanner class, better
  organization pushing it all to there instead of inner classes.

- More switching behavior encapsulated in GemsToUpdate, unlocking_all no
  longer needed between ConservativeDefinition and ConservativeResolver.