First attempt merging patch/update functionality.
Tests don't pass, but I think the integration scanner specs are now what
I may want them to be.

The differences between patching and updating in 0.6.x don't _quite_ align
in any way that's too predictable, and the failing specs in this commit
show why, I think.

Still need to tease this out - but I think next step is remove the
@patching flag and whatever behavior is there for 'patching a gem'
becomes part of strict mode. At least try that, see if it sticks.