ConservativeResolver results cached.
I disabled this a long time ago because I noticed that it broke real
life cases and am just now getting back to investigating the problem.
The fix is to dup the result out of the cache. The dup is important,
because in weird (large) cases Bundler will empty the result array
corrupting the cache. Bundler itself doesn't have this problem because
the super search_for does a select on its cached search results,
effectively duping it.

Had to do a lot of debug outputting to flush this out from a real life
case.

I also cleaned up the existing debug output, since the defensive coding
if no res or if an error thrown hasn't happened in real life for quite
some time, and those lines were rather noisy.