Re-enabling the fixup_empty_remotes hack.
I first encountered this in the midst of 0.5.x, 0.6.x work - one of many
"all the tests pass! -- it does nothing in real life" experiences. This
one with sidekiq and sidekiq-pro was particularly nasty, and I thought
could be a broader problem.

Then I found the `'update'=>true` problem in aace2142, and that fixed
_so many real problems_, I wasn't sure if `fixup_empty_remotes` was
still necessary. But I just tried it on our repo that uses sidekiq-pro
without the hack in, and it did have a similar failure to resolve
things. I put the hack back in, and it worked great. So, it's back for
g00d for now. At least until I shore up things with Bundler 1.11.x and
try it all again.