Fix to support Bundler 1.11.x
Fortunately, not much to do. Somehow, 1.10.x never cared about missing
dependencies in the integration tests with versions of addressable that
had runtime dependencies on rake and rspec. That was fixed in later
addressable versions, so I found a pair to swap it out with and
everything is jake.

I need to get some CI going that will test against both 1.10 and 1.11 of
Bundler.