Debugging fix.
The result of the super search_for is now dumped only on our flag, not
bundler's, and is only dumped when our code is kicked in (not cache
result returned). The way it was can be helpful early here in dev when
there's more questionable things afoot, but it's way too noisy otherwise
and shouldn't be triggered by bundler's resolver flag.