Add failing integration tests for Bundler 1.13
The integration into Bundler does not support target new version (which
vulnerabilities, in combination with `--prefer_minimal` rely on, and
vulns requiring jumping to a minor version when not allowed relies on,
too.

Need to decide what to do in these cases, but they're failing now. There
was nothing before highlighting these cases.