Vulnerable w/ minor up partially passing.
A few things moving here:

- Breaking out a separate test case for minimal and non-minimal
vulnerable upgrade requiring minor when minor isn't passed. Minimal case
is now working.

- Some refactoring a la what was done in Bundler proper to put
meaningful names on some of the comparisons.

- A change to get rid of `move_version_to_end` which was always a hack
of sorts. We're already sorting things, but as this criteria was even
_moar_ complicated, I just punted on working it into the case statement
and tacked it onto the end. But now I'm actually being screwed by it in
the Bundler proper code, and can't really fix the "non-minimal
vulnerable upgrade requiring minor when minor option isn't in use" case.
Now that I've shown how to do it here, I can update it in the bundler
code itself, which would have been merged already, except for weird test
failures still happening on master as a result, and limited attention to
be put on just this right now.