Add BUNDLER doc to document the undocumented.
As it says, Starting with 1.13.0.rc.2, a subset of bundler-patch
behavior was ported to Bundler itself. The plan is to leave it
undocumented and unsupported in 1.13 to give it a chance to be used and
flush out bugs and iron out some design decisions.

Part of that work before 1.14 (or maybe a later version) will be to
properly document `bundle update` as well as the [Conservative
Updating](http://bundler.io/v1.12/man/bundle-install.1.html#CONSERVATIVE-UPDATING)
section of `bundle install`, but right now we need a placeholder
document showing the new stuff so folks know how to use it.

This is largely copy/pasta of the bundler-patch [README](README.md).