Ensure cache is updated after install.
Fixes #36. In cases where vendor/cache exists, we need to make sure it
gets updated.

The underlying problem here is Bundler doesn't have a reusable "and now
install the gems and do all the other stuff that should be done" bit -
it's duped throughout the CLI classes. Maybe that'll happen in the
future.