Documentation updates.
BUNDLER is gone now that README can link over to actual Bundler.io
documentation. Lots of fixups in README.

Added a new DEBUG_PATCH_RESOLVER bit of data dumping the list of
unlocked gems. Note that the Bundler resolver itself can "lock" gems as
vertices in the Molinillo resolution process that won't necessarily
show in the "unlocked" list. The "unlocked" dump is more the gems
requested to be unlocked.