Test enhancement for different ruby test.
Refactored GemfileLockFixture to split out support for separate gemfile
and lockfile creation, since if we try to create the lockfile in the
separate Ruby case, it fails straight off in the setup of the
integration test.

Now I can continue on, trying to cobble together a Ruby bin path within
TargetBundle and unit specs there.