Add dir detection in Gemfile opt handling.
Fixes #47

Extract out CLI::Options, and ensure that if a target directory is
passed as the Gemfile option, that case is handled and we refer to the
default Gemfile in that dir. Otherwise it used to react as if nothing
needed to be done, when in fact it hadn't really done anything.