Fix for CI. Ensure GEM_HOME set.
GEM_HOME is now passed to the shelled out command. This was _only_ a
problem with CI and rvm that I wasn't seeing locally with rbenv. All
the paths are enclosed in quotes as well to help with that eventual
problem (spaces in paths).

A bug fix within the test code as well to make sure the output is still
returned to the test so its asserts will work.