Version 1.2.0 to support Bundler 2.
No production code changes to make this happen, but a lot of
accumulated build cruft had to be sorted out to get Travis happy after
a couple of years, plus a change to the bundler requirement in the
gemspec.

This is an empty commit with some explanations as to a weirdness in the
1.2.0 release. The tests are all failing right now, but mostly on two
integration tests that test using the target Ruby to execute
bundler-patch, which is in a catch-22 of needing a released version of
bundler-patch which supports a dependency requirement of Bundler 2.

These pass locally, but can't pass on Travis until the gem is released.

There are _also_ a few failing cases with Bundler 2 that may or may not
be real bugs.

Since I'm not making anything _worse_ to my knowledge, and not many
folks use this gem anyway, I'm forging ahead with the release of 1.2.0
knowing some additional releases may be hot on its heels.